+++ cookbooks/arcgis-notebooks/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file is used to list changes made in each version of the arcgis-notebooks cookbook.
 
-# 3.5.0
+# 3.6.0
+-------
 
-Initial release of the cookbook.
+- Added support for ArcGIS Notebook Server 10.8.1.
+- Added unregister_machine recipe.
+
+# 3.5.0
+-------
+
+- Initial release of the cookbook.